RecipeSummaryScreen refactored

+++ app/src/main/java/com/aamo/cookbook/features/recipe/form/models/RecipeFormIngredientFields.kt
@@ -6,6 +6,6 @@ import java.util.UUID
 data class RecipeFormIngredientFields(
   val uuid: UUID = UUID.randomUUID(),
   val name: String = String.EMPTY,
-  val amount: Float? = null,
+  val amount: Double? = null,
   val unit: String = String.EMPTY,
 )
