+++ app/src/test/java/com/aamo/cookbook/test_utility/RecipeMocker.kt
@@ -141,7 +141,7 @@ class StepMocker(seed: Int = 0) {
 
 class IngredientMocker(seed: Int = 0) {
   private var ingredient = Ingredient(
-    id = 0, stepId = 0, name = "$seed", amount = seed.toFloat(), unit = "$seed"
+    id = 0, stepId = 0, name = "$seed", amount = seed.toDouble(), unit = "$seed"
   )
 
   fun modify(ingredient: (Ingredient) -> Ingredient): IngredientMocker {
